requirements.txt
ffmpeg-python
psycopg2-binary
langchain
